Fixed some small bugs here and there
- The division in AverageAuctionsPerParcel measure was in int with very bad things as a result
- Forgot to override method in ExponentialBackoffSLackReAuctionableParcel which resulted in wrong getCreator() call

+++ src/main/java/common/results/measures/BasicMeasure.java
@@ -152,7 +152,7 @@ public abstract class BasicMeasure<E> extends Measure<E> {
                 total += par.getNumberReAuctions();
             }
 
-            return Collections.singletonList(Double.toString(total / getParcelsFromRun(result).size()));
+            return Collections.singletonList(Double.toString((float) total / getParcelsFromRun(result).size()));
         }
     }
 }
